requirements.txt
python-dotenv
requests
schedule
selenium
webdriver-manager
psutil
pyyaml